+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Pre-release
 
+### Implemented enhancements
+
+- [TT-1328] Added Investment Support Directory profile.
+- [TT-1352] Added Investment Support Directory search.
+
 ### Fixed bugs:
 
 - [[CMS-1256]](https://uktrade.atlassian.net/browse/CMS-1256) Fix vertical spacing in footer
@@ -21,8 +26,6 @@
 - Hard-code /trade/ url prefix (was previously feature flagged).
 - Upgraded [Constants][directory-constants].
 - Removed application-level IP filtering (replaced with routing solution at platform level).
-- [TT-1328] Added Investment Support Directory profile.
-- [TT-1352] Added Investment Support Directory search.
 - Added fallback cache for API
 - GA360 changes FAB search to use q instead of term. Redirect also in place for backwards compatibility.