+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ### Fixed bugs:
 
-- [[CMS-1256]](https://uktrade.atlassian.net/browse/CMS-1256) Fix vertical spacing in footer
+- [[CMS-1256]](https://uktrade.atlassian.net/browse/CMS-1256) Fix mobile vertical spacing in footer
+- [[CMS-1262]](https://uktrade.atlassian.net/browse/CMS-1262) Fix mobile column spacing on industries list page
 
 ## [2019.04.25](https://github.com/uktrade/directory-ui-supplier/releases/tag/2019.04.25)